MXS-3362: Store sescmd history in client protocol
The code used for handling prepared statements can be repurposed to also
handle session commands. The only additional thing that needs to be done
is to store the query while we're waiting for it to give us a result.

The history itself is in the shared MYSQL_session object which is
accessible to all backend connections. This simplifies the replaying
process into writing each of the stored buffers into the outbound socket
once authentication has completed and waiting for all the results to
arrive.