MXS-3362: Remove unused readwritesplit code
The code is no longer used and can be removed.

The act of discarding connections whose responses don't match the expected
one was also removed. This will be either done by the protocol module or
it will be rewritten for the router. Either way, the old code can be
temporarily removed.