MXS-3362: Pause writes until PS finishes
Pausing the writes while a backend executes a prepared statement makes
sure that it knows of all prepared IDs. It is possible that, from the
viewpoint of the backend protocol, a client (in this case the router)
sends a COM_STMT_PREPARE and COM_STMT_CLOSE at the same time, the latter
using the ID that would be generated by the former. This is only possible
if a faster backend delivers the response to the COM_STMT_PREPARE which
allows the client to send more queries.

This pause doesn't need to happen during history replay as we already
queue new writes until the history replay has completed.

In addition, the delayed packet writing code needs to not rely on
m_delayed_packets being constant for the duration of the call. With the
addition of a separate state for COM_STMT_PREPARE handling, it's possible
(albeit unlikely) that multiple COM_STMT_PREPARE statements are sent at
the same time.