MXS-3362: Check responses to session commands
The backend protocol module now checks whether the response that it just
sent upstream is the same as the accepted response that is sent to the
client and if it's not it closes the connection.

For responses that arrive too early, the check is postponed until the
accepted response arrives. The client protocol then invokes a callback
once the correct response arrives.

One limitation of this approach is that when history pruning is
implemented, it is theoretically possible that a backend is so slow that
it doesn't get to execute the session command before it is removed from
the history. If this happens, the code currently assumes that the response
is correct. Another solution would be to close the connection with the
assumption that something is wrong with it. This would make sure that the
state is consistent even if this case.